refactor: split BookService into BookFileService and BookMetadataManageService
