+++ booklore-ui/src/app/features/book/service/book-patch.service.ts
@@ -6,7 +6,7 @@ import {Book, BookFileProgress, ReadStatus} from '../model/book.model';
 import {BookStateService} from './book-state.service';
 import {API_CONFIG} from '../../../core/config/api-config';
 import {ResetProgressType, ResetProgressTypes} from '../../../shared/constants/reset-progress-type';
-import {BookStatusUpdateResponse, PersonalRatingUpdateResponse} from './book.service';
+import {BookStatusUpdateResponse, PersonalRatingUpdateResponse} from '../model/book.model';
 
 @Injectable({
   providedIn: 'root',
