+++ booklore-ui/src/app/features/book/service/book.service.ts
@@ -2,31 +2,18 @@ import {inject, Injectable} from '@angular/core';
 import {first, from, Observable, of, throwError} from 'rxjs';
 import {HttpClient, HttpParams} from '@angular/common/http';
 import {catchError, distinctUntilChanged, filter, finalize, map, shareReplay, switchMap, tap} from 'rxjs/operators';
-import {AdditionalFile, AdditionalFileType, Book, BookDeletionResponse, BookMetadata, BookRecommendation, BookSetting, BookSyncResponse, BookType, BulkMetadataUpdateRequest, CreatePhysicalBookRequest, MetadataUpdateWrapper, ReadStatus} from '../model/book.model';
+import {Book, BookDeletionResponse, BookMetadata, BookRecommendation, BookSetting, BookStatusUpdateResponse, BookSyncResponse, BookType, CreatePhysicalBookRequest, PersonalRatingUpdateResponse, ReadStatus} from '../model/book.model';
 import {BookState} from '../model/state/book-state.model';
 import {API_CONFIG} from '../../../core/config/api-config';
 import {MessageService} from 'primeng/api';
 import {ResetProgressType} from '../../../shared/constants/reset-progress-type';
 import {AuthService} from '../../../shared/service/auth.service';
-import {FileDownloadService} from '../../../shared/service/file-download.service';
 import {Router} from '@angular/router';
 import {BookStateService} from './book-state.service';
 import {BookSocketService} from './book-socket.service';
 import {BookPatchService} from './book-patch.service';
 import {BookCacheService} from './book-cache.service';
 import {TranslocoService} from '@jsverse/transloco';
-
-export interface BookStatusUpdateResponse {
-  bookId: number;
-  readStatus: ReadStatus;
-  readStatusModifiedTime: string;
-  dateFinished?: string;
-}
-
-export interface PersonalRatingUpdateResponse {
-  bookId: number;
-  personalRating?: number;
-}
 
 @Injectable({
   providedIn: 'root',
@@ -38,7 +25,6 @@ export class BookService {
   private http = inject(HttpClient);
   private messageService = inject(MessageService);
   private authService = inject(AuthService);
-  private fileDownloadService = inject(FileDownloadService);
   private router = inject(Router);
   private bookStateService = inject(BookStateService);
   private bookSocketService = inject(BookSocketService);
@@ -449,199 +435,6 @@ export class BookService {
     return this.http.put<void>(`${this.url}/${bookId}/viewer-setting`, bookSetting);
   }
 
-  /*------------------ File Operations ------------------*/
-
-  getFileContent(bookId: number, bookType?: string): Observable<Blob> {
-    let url = `${this.url}/${bookId}/content`;
-    if (bookType) {
-      url += `?bookType=${bookType}`;
-    }
-    return this.http.get<Blob>(url, {responseType: 'blob' as 'json'});
-  }
-
-  downloadFile(book: Book): void {
-    const downloadUrl = `${this.url}/${book.id}/download`;
-    this.fileDownloadService.downloadFile(downloadUrl, book.primaryFile?.fileName ?? 'book');
-  }
-
-  downloadAllFiles(book: Book): void {
-    const downloadUrl = `${this.url}/${book.id}/download-all`;
-    const filename = book.metadata?.title
-      ? `${book.metadata.title.replace(/[^a-zA-Z0-9\-_]/g, '_')}.zip`
-      : `book-${book.id}.zip`;
-    this.fileDownloadService.downloadFile(downloadUrl, filename);
-  }
-
-  deleteAdditionalFile(bookId: number, fileId: number): Observable<void> {
-    const deleteUrl = `${this.url}/${bookId}/files/${fileId}`;
-    return this.http.delete<void>(deleteUrl).pipe(
-      tap(() => {
-        const currentState = this.bookStateService.getCurrentBookState();
-        const updatedBooks = (currentState.books || []).map(book => {
-          if (book.id === bookId) {
-            return {
-              ...book,
-              alternativeFormats: book.alternativeFormats?.filter(file => file.id !== fileId),
-              supplementaryFiles: book.supplementaryFiles?.filter(file => file.id !== fileId)
-            };
-          }
-          return book;
-        });
-
-        this.bookStateService.updateBookState({
-          ...currentState,
-          books: updatedBooks
-        });
-
-        this.messageService.add({
-          severity: 'success',
-          summary: this.t.translate('book.bookService.toast.fileDeletedSummary'),
-          detail: this.t.translate('book.bookService.toast.additionalFileDeletedDetail')
-        });
-      }),
-      catchError(error => {
-        this.messageService.add({
-          severity: 'error',
-          summary: this.t.translate('book.bookService.toast.fileDeleteFailedSummary'),
-          detail: error?.error?.message || error?.message || this.t.translate('book.bookService.toast.fileDeleteFailedDetail')
-        });
-        return throwError(() => error);
-      })
-    );
-  }
-
-  deleteBookFile(bookId: number, fileId: number, isPrimary: boolean): Observable<void> {
-    const deleteUrl = `${this.url}/${bookId}/files/${fileId}`;
-    return this.http.delete<void>(deleteUrl).pipe(
-      tap(() => {
-        const currentState = this.bookStateService.getCurrentBookState();
-        const updatedBooks = (currentState.books || []).map(book => {
-          if (book.id === bookId) {
-            if (isPrimary) {
-              // Primary file was deleted - promote first alternative to primary, or set null
-              const remainingAlternatives = book.alternativeFormats?.filter(file => file.id !== fileId) || [];
-              if (remainingAlternatives.length > 0) {
-                const [newPrimary, ...restAlternatives] = remainingAlternatives;
-                return {
-                  ...book,
-                  primaryFile: newPrimary,
-                  alternativeFormats: restAlternatives
-                };
-              } else {
-                return {
-                  ...book,
-                  primaryFile: undefined,
-                  alternativeFormats: []
-                };
-              }
-            } else {
-              // Alternative file was deleted
-              return {
-                ...book,
-                alternativeFormats: book.alternativeFormats?.filter(file => file.id !== fileId)
-              };
-            }
-          }
-          return book;
-        });
-
-        this.bookStateService.updateBookState({
-          ...currentState,
-          books: updatedBooks
-        });
-
-        this.messageService.add({
-          severity: 'success',
-          summary: this.t.translate('book.bookService.toast.fileDeletedSummary'),
-          detail: this.t.translate('book.bookService.toast.bookFileDeletedDetail')
-        });
-      }),
-      catchError(error => {
-        this.messageService.add({
-          severity: 'error',
-          summary: this.t.translate('book.bookService.toast.fileDeleteFailedSummary'),
-          detail: error?.error?.message || error?.message || this.t.translate('book.bookService.toast.fileDeleteFailedDetail')
-        });
-        return throwError(() => error);
-      })
-    );
-  }
-
-  uploadAdditionalFile(bookId: number, file: File, fileType: AdditionalFileType, description?: string): Observable<AdditionalFile> {
-    const formData = new FormData();
-    formData.append('file', file);
-
-    const isBook = fileType === AdditionalFileType.ALTERNATIVE_FORMAT;
-    formData.append('isBook', String(isBook));
-
-    if (isBook) {
-      const lower = (file?.name || '').toLowerCase();
-      const ext = lower.includes('.') ? lower.substring(lower.lastIndexOf('.') + 1) : '';
-      const bookType = ext === 'pdf'
-        ? 'PDF'
-        : ext === 'epub'
-          ? 'EPUB'
-          : (ext === 'cbz' || ext === 'cbr' || ext === 'cb7' || ext === 'cbt')
-            ? 'CBX'
-            : (ext === 'm4b' || ext === 'm4a' || ext === 'mp3')
-              ? 'AUDIOBOOK'
-              : null;
-
-      if (bookType) {
-        formData.append('bookType', bookType);
-      }
-    }
-    if (description) {
-      formData.append('description', description);
-    }
-
-    return this.http.post<AdditionalFile>(`${this.url}/${bookId}/files`, formData).pipe(
-      tap((newFile) => {
-        const currentState = this.bookStateService.getCurrentBookState();
-        const updatedBooks = (currentState.books || []).map(book => {
-          if (book.id === bookId) {
-            const updatedBook = {...book};
-            if (fileType === AdditionalFileType.ALTERNATIVE_FORMAT) {
-              updatedBook.alternativeFormats = [...(book.alternativeFormats || []), newFile];
-            } else {
-              updatedBook.supplementaryFiles = [...(book.supplementaryFiles || []), newFile];
-            }
-            return updatedBook;
-          }
-          return book;
-        });
-
-        this.bookStateService.updateBookState({
-          ...currentState,
-          books: updatedBooks
-        });
-
-        this.messageService.add({
-          severity: 'success',
-          summary: this.t.translate('book.bookService.toast.fileUploadedSummary'),
-          detail: this.t.translate('book.bookService.toast.fileUploadedDetail')
-        });
-      }),
-      catchError(error => {
-        this.messageService.add({
-          severity: 'error',
-          summary: this.t.translate('book.bookService.toast.uploadFailedSummary'),
-          detail: error?.error?.message || error?.message || this.t.translate('book.bookService.toast.uploadFailedDetail')
-        });
-        return throwError(() => error);
-      })
-    );
-  }
-
-  downloadAdditionalFile(book: Book, fileId: number): void {
-    const additionalFile = [
-      ...(book.alternativeFormats || []),
-      ...(book.supplementaryFiles || [])
-    ].find((f: AdditionalFile) => f.id === fileId);
-    const downloadUrl = `${this.url}/${book.id}/files/${fileId}/download`;
-    this.fileDownloadService.downloadFile(downloadUrl, additionalFile?.fileName ?? 'file');
-  }
-
   /*------------------ Progress & Status Tracking ------------------*/
 
   updateLastReadTime(bookId: number): void {
@@ -682,175 +475,6 @@ export class BookService {
     return this.bookPatchService.updatePersonalRating(bookIds, rating);
   }
 
-  /*------------------ Metadata Operations ------------------*/
-
-
-  updateBookMetadata(bookId: number | undefined, wrapper: MetadataUpdateWrapper, mergeCategories: boolean): Observable<BookMetadata> {
-    const params = new HttpParams().set('mergeCategories', mergeCategories.toString());
-    return this.http.put<BookMetadata>(`${this.url}/${bookId}/metadata`, wrapper, {params}).pipe(
-      map(updatedMetadata => {
-        this.handleBookMetadataUpdate(bookId!, updatedMetadata);
-        return updatedMetadata;
-      })
-    );
-  }
-
-  updateBooksMetadata(request: BulkMetadataUpdateRequest): Observable<void> {
-    return this.http.put(`${this.url}/bulk-edit-metadata`, request).pipe(
-      map(() => void 0)
-    );
-  }
-
-  toggleAllLock(bookIds: Set<number>, lock: string): Observable<void> {
-    const requestBody = {
-      bookIds: Array.from(bookIds),
-      lock: lock
-    };
-    return this.http.put<BookMetadata[]>(`${this.url}/metadata/toggle-all-lock`, requestBody).pipe(
-      tap((updatedMetadataList) => {
-        const currentState = this.bookStateService.getCurrentBookState();
-        const updatedBooks = (currentState.books || []).map(book => {
-          const updatedMetadata = updatedMetadataList.find(meta => meta.bookId === book.id);
-          return updatedMetadata ? {...book, metadata: updatedMetadata} : book;
-        });
-        this.bookStateService.updateBookState({...currentState, books: updatedBooks});
-      }),
-      map(() => void 0),
-      catchError((error) => {
-        throw error;
-      })
-    );
-  }
-
-  toggleFieldLocks(bookIds: number[] | Set<number>, fieldActions: Record<string, 'LOCK' | 'UNLOCK'>): Observable<void> {
-    const bookIdSet = bookIds instanceof Set ? bookIds : new Set(bookIds);
-
-    const requestBody = {
-      bookIds: Array.from(bookIdSet),
-      fieldActions
-    };
-
-    return this.http.put<void>(`${this.url}/metadata/toggle-field-locks`, requestBody).pipe(
-      tap(() => {
-        const currentState = this.bookStateService.getCurrentBookState();
-        const updatedBooks = (currentState.books || []).map(book => {
-          if (!bookIdSet.has(book.id)) return book;
-          const updatedMetadata = {...book.metadata};
-          for (const [field, action] of Object.entries(fieldActions)) {
-            const lockField = field.endsWith('Locked') ? field : `${field}Locked`;
-            if (lockField in updatedMetadata) {
-              (updatedMetadata as Record<string, unknown>)[lockField] = action === 'LOCK';
-            }
-          }
-          return {
-            ...book,
-            metadata: updatedMetadata
-          };
-        });
-        this.bookStateService.updateBookState({
-          ...currentState,
-          books: updatedBooks as Book[]
-        });
-      }),
-      catchError(error => {
-        this.messageService.add({
-          severity: 'error',
-          summary: this.t.translate('book.bookService.toast.fieldLockFailedSummary'),
-          detail: this.t.translate('book.bookService.toast.fieldLockFailedDetail'),
-        });
-        throw error;
-      })
-    );
-  }
-
-  consolidateMetadata(metadataType: 'authors' | 'categories' | 'moods' | 'tags' | 'series' | 'publishers' | 'languages', targetValues: string[], valuesToMerge: string[]): Observable<unknown> {
-    const payload = {metadataType, targetValues, valuesToMerge};
-    return this.http.post(`${this.url}/metadata/manage/consolidate`, payload).pipe(
-      tap(() => {
-        this.refreshBooks();
-      })
-    );
-  }
-
-  deleteMetadata(metadataType: 'authors' | 'categories' | 'moods' | 'tags' | 'series' | 'publishers' | 'languages', valuesToDelete: string[]): Observable<unknown> {
-    const payload = {metadataType, valuesToDelete};
-    return this.http.post(`${this.url}/metadata/manage/delete`, payload).pipe(
-      tap(() => {
-        this.refreshBooks();
-      })
-    );
-  }
-
-  /*------------------ Cover Operations ------------------*/
-
-  getUploadCoverUrl(bookId: number): string {
-    return this.url + '/' + bookId + "/metadata/cover/upload"
-  }
-
-  uploadCoverFromUrl(bookId: number, url: string): Observable<BookMetadata> {
-    return this.http.post<BookMetadata>(`${this.url}/${bookId}/metadata/cover/from-url`, {url});
-  }
-
-  regenerateCovers(): Observable<void> {
-    return this.http.post<void>(`${this.url}/regenerate-covers`, {});
-  }
-
-  regenerateCover(bookId: number): Observable<void> {
-    return this.http.post<void>(`${this.url}/${bookId}/regenerate-cover`, {});
-  }
-
-  getFileMetadata(bookId: number): Observable<BookMetadata> {
-    return this.http.get<BookMetadata>(`${this.url}/${bookId}/file-metadata`);
-  }
-
-  generateCustomCover(bookId: number): Observable<void> {
-    return this.http.post<void>(`${this.url}/${bookId}/generate-custom-cover`, {});
-  }
-
-  generateCustomCoversForBooks(bookIds: number[]): Observable<void> {
-    return this.http.post<void>(`${this.url}/bulk-generate-custom-covers`, {bookIds});
-  }
-
-  regenerateCoversForBooks(bookIds: number[]): Observable<void> {
-    return this.http.post<void>(`${this.url}/bulk-regenerate-covers`, {bookIds});
-  }
-
-  uploadAudiobookCoverFromUrl(bookId: number, url: string): Observable<BookMetadata> {
-    return this.http.post<BookMetadata>(`${this.url}/${bookId}/metadata/audiobook-cover/from-url`, {url});
-  }
-
-  uploadAudiobookCoverFromFile(bookId: number, file: File): Observable<void> {
-    const formData = new FormData();
-    formData.append('file', file);
-    return this.http.post<void>(`${this.url}/${bookId}/metadata/audiobook-cover/upload`, formData);
-  }
-
-  getUploadAudiobookCoverUrl(bookId: number): string {
-    return this.url + '/' + bookId + "/metadata/audiobook-cover/upload";
-  }
-
-  regenerateAudiobookCover(bookId: number): Observable<void> {
-    return this.http.post<void>(`${this.url}/${bookId}/regenerate-audiobook-cover`, {});
-  }
-
-  generateCustomAudiobookCover(bookId: number): Observable<void> {
-    return this.http.post<void>(`${this.url}/${bookId}/generate-custom-audiobook-cover`, {});
-  }
-
-  supportsDualCovers(book: Book): boolean {
-    const allFiles = [book.primaryFile, ...(book.alternativeFormats || [])].filter(f => f?.bookType);
-    const hasAudiobook = allFiles.some(f => f!.bookType === 'AUDIOBOOK');
-    const hasEbook = allFiles.some(f => f!.bookType !== 'AUDIOBOOK');
-    return hasAudiobook && hasEbook;
-  }
-
-  bulkUploadCover(bookIds: number[], file: File): Observable<void> {
-    const formData = new FormData();
-    formData.append('file', file);
-    formData.append('bookIds', bookIds.join(','));
-    return this.http.post<void>(`${this.url}/bulk-upload-cover`, formData);
-  }
-
   /*------------------ Websocket Handlers ------------------*/
 
   handleNewlyCreatedBook(book: Book): void {
@@ -875,47 +499,5 @@ export class BookService {
 
   handleMultipleBookCoverPatches(patches: { id: number; coverUpdatedOn: string }[]): void {
     this.bookSocketService.handleMultipleBookCoverPatches(patches);
-  }
-
-  /*------------------ Book File Attachment ------------------*/
-
-  attachBookFiles(targetBookId: number, sourceBookIds: number[], deleteSourceBooks: boolean): Observable<Book> {
-    return this.http.post<Book>(`${this.url}/${targetBookId}/attach-file`, {
-      sourceBookIds,
-      deleteSourceBooks
-    }).pipe(
-      tap(updatedBook => {
-        const currentState = this.bookStateService.getCurrentBookState();
-        let updatedBooks = (currentState.books || []).map(book =>
-          book.id === targetBookId ? updatedBook : book
-        );
-
-        // If deleteSourceBooks, remove source books from state
-        if (deleteSourceBooks) {
-          const sourceIdSet = new Set(sourceBookIds);
-          updatedBooks = updatedBooks.filter(book => !sourceIdSet.has(book.id));
-        }
-
-        this.bookStateService.updateBookState({
-          ...currentState,
-          books: updatedBooks
-        });
-
-        const fileCount = sourceBookIds.length;
-        this.messageService.add({
-          severity: 'success',
-          summary: this.t.translate('book.bookService.toast.filesAttachedSummary'),
-          detail: this.t.translate('book.bookService.toast.filesAttachedDetail', {count: fileCount})
-        });
-      }),
-      catchError(error => {
-        this.messageService.add({
-          severity: 'error',
-          summary: this.t.translate('book.bookService.toast.attachmentFailedSummary'),
-          detail: error?.error?.message || error?.message || this.t.translate('book.bookService.toast.attachmentFailedDetail')
-        });
-        return throwError(() => error);
-      })
-    );
   }
 }
