+++ booklore-ui/src/app/features/metadata/component/book-metadata-center/metadata-viewer/metadata-viewer.component.ts
@@ -3,6 +3,7 @@ import {Button} from 'primeng/button';
 import {AsyncPipe, DecimalPipe, NgClass} from '@angular/common';
 import {combineLatest, Observable} from 'rxjs';
 import {BookService} from '../../../../book/service/book.service';
+import {BookFileService} from '../../../../book/service/book-file.service';
 import {Rating, RatingRateEvent} from 'primeng/rating';
 import {FormsModule} from '@angular/forms';
 import {Book, BookFile, BookMetadata, BookRecommendation, BookType, ComicMetadata, FileInfo, ReadStatus} from '../../../../book/model/book.model';
@@ -54,6 +55,7 @@ export class MetadataViewerComponent implements OnInit, OnChanges {
   private emailService = inject(EmailService);
   private messageService = inject(MessageService);
   private bookService = inject(BookService);
+  private bookFileService = inject(BookFileService);
   private taskHelperService = inject(TaskHelperService);
   protected urlHelper = inject(UrlHelperService);
   protected userService = inject(UserService);
@@ -484,11 +486,11 @@ export class MetadataViewerComponent implements OnInit, OnChanges {
   }
 
   download(book: Book) {
-    this.bookService.downloadFile(book);
+    this.bookFileService.downloadFile(book);
   }
 
   downloadAdditionalFile(book: Book, fileId: number) {
-    this.bookService.downloadAdditionalFile(book, fileId);
+    this.bookFileService.downloadAdditionalFile(book, fileId);
   }
 
   // Event handlers for MetadataTabsComponent
@@ -505,7 +507,7 @@ export class MetadataViewerComponent implements OnInit, OnChanges {
   }
 
   onDownloadAllFiles(event: DownloadAllFilesEvent): void {
-    this.bookService.downloadAllFiles(event.book);
+    this.bookFileService.downloadAllFiles(event.book);
   }
 
   onDeleteBookFile(event: DeleteBookFileEvent): void {
@@ -526,7 +528,7 @@ export class MetadataViewerComponent implements OnInit, OnChanges {
       rejectButtonStyleClass: 'p-button-secondary',
       acceptButtonStyleClass: 'p-button-danger',
       accept: () => {
-        this.bookService.deleteAdditionalFile(bookId, fileId).subscribe({
+        this.bookFileService.deleteAdditionalFile(bookId, fileId).subscribe({
           next: () => {
             this.messageService.add({
               severity: 'success',
@@ -572,7 +574,7 @@ export class MetadataViewerComponent implements OnInit, OnChanges {
       rejectButtonStyleClass: 'p-button-secondary',
       acceptButtonStyleClass: 'p-button-danger',
       accept: () => {
-        this.bookService.deleteBookFile(book.id, fileId, isPrimary).subscribe({
+        this.bookFileService.deleteBookFile(book.id, fileId, isPrimary).subscribe({
           next: () => {
             this.messageService.add({
               severity: 'success',
